deletion control

+++ ServerLibrary/DataAccess.cs
@@ -180,7 +180,7 @@ namespace ServerLibrary
 
 
 
-        public static void removefromCanal(string canalName, string username)
+        public static void removefromCanal(string canalName, string username, User user)
         {
             using (IDbConnection databaseConnection = new SQLiteConnection(LoadConnectionString()))
             {
@@ -189,8 +189,9 @@ namespace ServerLibrary
                 if (result != null)
                 {
                     var check = databaseConnection.QuerySingleOrDefault(string.Format("SELECT * FROM {0} WHERE username = \"{1}\"", canalName, username));
+                    var admincheck =  databaseConnection.QuerySingleOrDefault(string.Format("SELECT * FROM {0} WHERE username = \"{1}\" AND administrator = 1", canalName, user.Name));
 
-                    if (check != null)
+                    if (check != null && admincheck != null)
                     {
                         databaseConnection.Execute(string.Format("DELETE FROM {0} WHERE username = \"{1}\"",canalName,username));
                     }
@@ -208,9 +209,23 @@ namespace ServerLibrary
                 }
             }
         }
-        public static void leaveCanal(string canalname,User user)
+        public static void leaveCanal(string canalName,User user)
         {
-            removefromCanal(canalname, user.Name);
+          
+           using (IDbConnection databaseConnection = new SQLiteConnection(LoadConnectionString()))
+            {
+                var result = databaseConnection.QuerySingleOrDefault(string.Format("SELECT * FROM canals WHERE name = \"{0}\"", canalName));
+
+                if (result != null)
+                {
+                    var check = databaseConnection.QuerySingleOrDefault(string.Format("SELECT * FROM {0} WHERE username = \"{1}\"", canalName, user.Name));
+
+                    if (check != null)
+                    {
+                        databaseConnection.Execute(string.Format("DELETE FROM {0} WHERE username = \"{1}\"",canalName,user.Name));
+                    }
+                }
+            }
         }
 
         public static string[] listuserCanal(string canalName)
